pyproject.toml
[project]
name = "preforge"
version = "0.1.0"
description = "각종 문서(docx, pptx, xlsx, html, md 등)를 읽거나 변환할 수 있는 파이썬 라이브러리. NER 모델과 AI Agent를 활용한 문서 분석 기능 제공."
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "dotnetpower" }
]
keywords = ["document", "parsing", "ner", "ai-agent", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chardet>=5.2.0",
    "lxml>=6.0.2",
    "markdown>=3.10",
    "opencv-python>=4.13.0.90",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.0",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "scikit-image>=0.26.0",
    "spacy>=3.8.11",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "flake8>=7.3.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
